Update tables.md
Added a beginner friendly example that doesn't require using responders. I specifically had a lot of trouble getting the styling (column widths) to apply while an element was being dragged using the current fixed layout strategy. The current guide makes it sound like you must use event responders and set styling based on if the element is dragging or not. This is not necessary (see sandbox link), and adds complexity for many beginners.